refactor: like 수 jpa 메서드로 가져오기

+++ src/main/kotlin/com/yapp/lettie/api/timecapsule/service/dto/TimeCapsuleDetailDto.kt
@@ -9,7 +9,7 @@ data class TimeCapsuleDetailDto(
     val subtitle: String?,
     val openAt: LocalDateTime,
     val participantCount: Int,
-    val likeCount: Int,
+    val likeCount: Long,
     val isLiked: Boolean? = null,
     val status: TimeCapsuleStatus,
     val remainingTime: RemainingTimeDto? = null,
